fix: cuda_simsafe owns the single conditional cupy/cupyx import; stream interop lives in mem_manager
cupy and cupyx are imported eagerly at package import on a real GPU
(with the install-instruction ImportError) and are None under the
CUDA simulator; consumers import them from cubie.cuda_simsafe.
_import_cupy and every function-local cupy/cupyx import are gone.
cupy_emm.py is deleted: with the EMM interface removed, its only
remaining members (_numba_stream_ptr, current_cupy_stream) belong to
their sole consumer, mem_manager. Stream-forwarding tests live in
test_memmgmt.py; docs and AGENTS.md describe the import rule.
