+++ docs/source/API_reference/memory/current_cupy_stream.rst
@@ -1,6 +1,6 @@
 current_cupy_stream
 ===================
 
-.. currentmodule:: cubie.memory.cupy_emm
+.. currentmodule:: cubie.memory.mem_manager
 
 .. autofunction:: current_cupy_stream
